refactor(onepassword): extract in-process port-forward into shared k8s util
Move the in-process Kubernetes port-forward transport out of onepassword/item.ts
into packages/sector7/k8s/port-forward.ts (withPortForward + PortForwardTarget),
so it can be shared with litellm/admin.ts once sector7#258 lands (ADR 031
follow-up). The function body is unchanged; item.ts now imports it.

The shared module keeps the serialization contract explicit: no top-level
runtime imports, only lazy await import() inside the function, so a dynamic
provider closure that references withPortForward still serializes. Verified by
running Pulumi's real serializeFunction against the tsc-compiled output (the
mock-based vitest suite can't — vite rewrites await import() into a this-
capturing module-runner call, the same test-env artifact sector7#258 documents).
